examples/python/adk/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Code for the x402<>A2A Payments Demo"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "google-adk",
    "x402>=0.1.4",
    "click>=8.2.0",
    "eth-account>=0.13.7",
    "google-adk>=0.5.0",
    "httpx>=0.28.1",
    "uvicorn>=0.34.2",
    "web3>=7.10.0",
    "cdp-sdk>=1.20.0",
    "a2a-x402"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
a2a_x402 = { path = "../../a2a_x402", editable = true } 